+++ lib/pages/splash/splash_page.dart
@@ -1,13 +1,14 @@
 import 'package:flutter/material.dart';
 import 'package:get/get.dart';
 import 'package:q_overlay/q_overlay.dart';
+
 import '../../helpers/constants.dart';
 import '../../services/storage_service.dart';
+import '../main/controllers/main_controller.dart';
+import '../main/views/main_view.dart';
 import '../settings/controllers/settings_controller.dart';
 import '../settings/models/models.dart';
 import '../settings/views/settings_view.dart';
-import '../main/controllers/main_controller.dart';
-import '../main/views/main_view.dart';
 
 class SplashPage extends StatefulWidget {
   const SplashPage({Key? key}) : super(key: key);
@@ -56,19 +57,22 @@ class _SplashPageState extends State<SplashPage> {
     _isPanelOpen = true;
     final app = await QPanel(
       child: const SettingsPage(isSelectApp: true),
-      backgroundFilter: BackgroundFilterSettings.transparent(),
+      backgroundFilter: const BackgroundFilterSettings(
+        blurX: 0.001,
+        blurY: 0.001,
+      ),
     ).show<AppLocalFile>();
     _isPanelOpen = false;
     if (app == null) {
       setState(() {});
       return;
     }
-    final loclas = await _storageController.loadLocals(app);
-    if (loclas == null) {
+    final locals = await _storageController.loadLocals(app);
+    if (locals == null) {
       setState(() {});
       return;
     }
-    Get.lazyPut(() => MainController(appfile: app, locals: loclas));
+    Get.lazyPut(() => MainController(appFile: app, locals: locals));
     Navigator.of(context)
         .pushReplacement(MaterialPageRoute(builder: (_) => const MainView()));
   }